// packages/xyd-opencli-remark/src/types.ts
export interface OpencliSpecJson {
  /**
   * The OpenCLI version number
   */
  opencli: string;
  info: CliInfo;
  conventions?: Conventions;
  /**
   * Root command arguments
   */
  arguments?: Argument[];
  /**
   * Root command options
   */
  options?: Option[];
  /**
   * Root command sub commands
   */
  commands?: Command[];
  /**
   * Root command exit codes
   */
  exitCodes?: ExitCode[];
  /**
   * Examples of how to use the CLI
   */
  examples?: string[];
  /**
   * Indicates whether or not the command requires interactive input
   */
  interactive?: boolean;
  /**
   * Custom metadata
   */
  metadata?: Metadata[];
  [k: string]: unknown;
}
/**
 * Information about the CLI
 */
export interface CliInfo {
  /**
   * The application title
   */
  title: string;
  /**
   * A short summary of the application
   */
  summary?: string;
  /**
   * A description of the application
   */
  description?: string;
  contact?: Contact;
  license?: License;
  /**
   * The application version
   */
  version: string;
  [k: string]: unknown;
}
/**
 * The contact information
 */
export interface Contact {
  /**
   * The identifying name of the contact person/organization
   */
  name?: string;
  /**
   * The URI for the contact information. This MUST be in the form of a URI.
   */
  url?: string;
  /**
   * The email address of the contact person/organization. This MUST be in the form of an email address.
   */
  email?: string;
  [k: string]: unknown;
}
/**
 * The application license
 */
export interface License {
  /**
   * The license name
   */
  name?: string;
  /**
   * The SPDX license identifier
   */
  identifier?: string;
  /**
   * The URI for the license. This MUST be in the form of a URI
   */
  url?: string;
  [k: string]: unknown;
}
/**
 * The conventions used by the CLI
 */
export interface Conventions {
  /**
   * Whether or not grouping of short options are allowed
   */
  groupOptions?: boolean;
  /**
   * The option argument separator
   */
  optionSeparator?: string;
  [k: string]: unknown;
}
export interface Argument {
  /**
   * The argument name
   */
  name: string;
  /**
   * Whether or not the argument is required
   */
  required?: boolean;
  arity?: Arity;
  /**
   * A list of accepted values
   */
  acceptedValues?: string[];
  /**
   * The argument group
   */
  group?: string;
  /**
   * The argument description
   */
  description?: string;
  /**
   * Whether or not the argument is hidden
   */
  hidden?: boolean;
  /**
   * Custom metadata
   */
  metadata?: Metadata[];
  [k: string]: unknown;
}
/**
 * The argument arity. Arity defines the minimum and maximum number of argument values
 */
export interface Arity {
  /**
   * The minimum number of values allowed
   */
  minimum?: number;
  /**
   * The maximum number of values allowed
   */
  maximum?: number;
  [k: string]: unknown;
}
export interface Metadata {
  name: string;
  value?: unknown;
  [k: string]: unknown;
}
export interface Option {
  /**
   * The option name
   */
  name: string;
  /**
   * Whether or not the option is required
   */
  required?: boolean;
  /**
   * The option's aliases
   */
  aliases?: string[];
  /**
   * The option's arguments
   */
  arguments?: Argument[];
  /**
   * The option group
   */
  group?: string;
  /**
   * The option description
   */
  description?: string;
  /**
   * Specifies whether the option is accessible from the immediate parent command and, recursively, from its subcommands
   */
  recursive?: boolean;
  /**
   * Whether or not the option is hidden
   */
  hidden?: boolean;
  /**
   * Custom metadata
   */
  metadata?: Metadata[];
  [k: string]: unknown;
}
export interface Command {
  /**
   * The command name
   */
  name: string;
  /**
   * The command aliases
   */
  aliases?: string[];
  /**
   * The command options
   */
  options?: Option[];
  /**
   * The command arguments
   */
  arguments?: Argument[];
  /**
   * The command's sub commands
   */
  commands?: Command[];
  /**
   * The command's exit codes
   */
  exitCodes?: ExitCode[];
  /**
   * The command description
   */
  description?: string;
  /**
   * Whether or not the command is hidden
   */
  hidden?: boolean;
  /**
   * Examples of how to use the command
   */
  examples?: string[];
  /**
   * Indicate whether or not the command requires interactive input
   */
  interactive?: boolean;
  /**
   * Custom metadata
   */
  metadata?: Metadata[];
  [k: string]: unknown;
}
export interface ExitCode {
  /**
   * The exit code
   */
  code: number;
  /**
   * The exit code description
   */
  description?: string;
  [k: string]: unknown;
}